Factor out common F680 wizard logic to parent class
