+++ exporter/f680/application_sections/views.py
@@ -1,0 +1,59 @@
+from http import HTTPStatus
+
+from django.shortcuts import redirect
+from django.urls import reverse
+
+from core.auth.views import LoginRequiredMixin
+from core.decorators import expect_status
+from core.wizard.views import BaseSessionWizardView
+
+from exporter.f680.services import patch_f680_application, get_f680_application
+from exporter.f680.payloads import F680PatchPayloadBuilder
+from exporter.f680.views import F680FeatureRequiredMixin
+
+
+class F680ApplicationSectionWizard(LoginRequiredMixin, F680FeatureRequiredMixin, BaseSessionWizardView):
+    form_list = []
+    condition_dict = {}
+    section = None
+
+    def setup(self, request, *args, **kwargs):
+        super().setup(request, *args, **kwargs)
+        self.application, _ = self.get_f680_application(kwargs["pk"])
+
+    @expect_status(
+        HTTPStatus.OK,
+        "Error retrieving F680 application",
+        "Unexpected error retrieving F680 application",
+        reraise_404=True,
+    )
+    def get_f680_application(self, pk):
+        return get_f680_application(self.request, pk)
+
+    @expect_status(
+        HTTPStatus.OK,
+        "Error updating F680 application",
+        "Unexpected error updating F680 application",
+    )
+    def patch_f680_application(self, data):
+        return patch_f680_application(self.request, self.application["id"], data)
+
+    def get_form_initial(self, step):
+        return self.application.get("application", {}).get(self.section, {}).get("answers", {})
+
+    def get_success_url(self, application_id):
+        return reverse(
+            "f680:summary",
+            kwargs={
+                "pk": application_id,
+            },
+        )
+
+    def get_payload(self, form_dict):
+        current_application = self.application.get("application", {})
+        return F680PatchPayloadBuilder().build(self.section, current_application, form_dict)
+
+    def done(self, form_list, form_dict, **kwargs):
+        data = self.get_payload(form_dict)
+        response_data, _ = self.patch_f680_application(data)
+        return redirect(self.get_success_url(response_data["id"]))
